+++ HISTORY.md
@@ -3,6 +3,7 @@
 ### 2014-07-31 - v5.3 - (Dev time: 1 hours) ###
 - Tweak: Tweak the names of the menu items (Issue #73)
 - Tweak: Add information about `DISABLE_WP_CRON` in system info
+- Tweak: Tweak labels in Scheduled Jobs page (Issue #71)
 
 
 ### 2014-07-03 - v5.2 - (Dev time: 8 hours) ###