+++ HISTORY.md
@@ -4,6 +4,7 @@
 - Tweak: Tweak the names of the menu items (Issue #73)
 - Tweak: Add information about `DISABLE_WP_CRON` in system info
 - Tweak: Tweak labels in Scheduled Jobs page (Issue #71)
+- Tweak: Removed unused variable
 
 - Bug: Fixed a warning that happened because of duplicate call to `add_meta_boxes`