+++ HISTORY.md
@@ -6,6 +6,7 @@
 - Tweak: Tweak labels in Scheduled Jobs page (Issue #71)
 - Tweak: Removed unused variable
 - Tweak: Removed old compatibility code for `The Events Calendar` plugin
+- Tweak: Add filters to extend menu items (Issue #74)
 
 - Bug: Fixed a warning that happened because of duplicate call to `add_meta_boxes`