+++ HISTORY.md
@@ -10,6 +10,7 @@
 - Tweak: Removed old compatibility code for `The Events Calendar` plugin
 - Tweak: Add filters to extend menu items (Issue #74)
 - Tweak: Add filters to extend meta boxes in each page (Issue #75)
+- Tweak: Remove `upgraded from` from system info (Issue #77)
 
 - Bug: Fixed a warning that happened because of duplicate call to `add_meta_boxes`