+++ HISTORY.md
@@ -1,7 +1,8 @@
 ## Changelog ##
 
-### 2014-08-13 - v5.3 - (Dev time: 10 hours) ###
+### 2014-08-13 - v5.3 - (Dev time: 15 hours) ###
 - New: Ability to delete Jetpack Contact Form Messages (Issue #72)
+- New: Settings screen for addons (Issue #78)
 
 - Tweak: Tweak the names of the menu items (Issue #73)
 - Tweak: Add information about `DISABLE_WP_CRON` in system info