+++ HISTORY.md
@@ -1,8 +1,9 @@
 ## Changelog ##
 
-### 2014-08-13 - v5.3 - (Dev time: 15 hours) ###
+### 2014-08-17 - v5.3 - (Dev time: 17 hours) ###
 - New: Ability to delete Jetpack Contact Form Messages (Issue #72)
 - New: Settings screen for addons (Issue #78)
+- New: Add setting helper functions for addons
 
 - Tweak: Tweak the names of the menu items (Issue #73)
 - Tweak: Add information about `DISABLE_WP_CRON` in system info